+++ src/Regressors/MLPRegressor.php
@@ -175,7 +175,7 @@ class MLPRegressor implements Estimator, Learner, Online, Verbose, Persistable
         float $alpha = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
-        int $window = 3,
+        int $window = 5,
         float $holdOut = 0.1,
         ?RegressionLoss $costFn = null,
         ?Metric $metric = null