+++ CHANGELOG.md
@@ -2,6 +2,8 @@
     - Change Token Hashing Vectorizer hashing function
     - Gradient Boost always uses mean base estimator
     - Move dummy estimators to the Extras package
+    - Increase default MLP window from 3 to 5
+    - Decrease default Gradient Boost window from 10 to 5
 
 - 1.2.0
     - Added Logit Boost classifier